SVCar/tools/requirements.txt
paho-mqtt>=2.1,<3
amqtt>=0.11,<0.12
